+++ src/gufo/thor/services/traefik.py
@@ -16,7 +16,7 @@ from typing import Optional
 
 # Gufo Thor modules
 from ..config import Config, ServiceConfig
-from .base import BaseService
+from .base import BaseService, ComposeDependsCondition
 from .consul import consul
 
 
@@ -26,6 +26,13 @@ class TraefikService(BaseService):
     name = "traefik"
     dependencies = (consul,)
     compose_image = "traefik:2.10"
+    compose_depends_condition = ComposeDependsCondition.HEALTHY
+    compose_healthcheck = {
+        "test": ["CMD", "traefik", "healthcheck", "--ping"],
+        "interval": "3s",
+        "timeout": "2s",
+        "retries": 3,
+    }
     compose_volumes = ["./etc/traefik/:/etc/traefik/:ro"]
 
     def prepare_compose_config(