+++ src/gufo/thor/targets/compose.py
@@ -43,36 +43,18 @@ class ComposeTarget(BaseTarget):
         etc = Path("etc")
         ensure_directory(etc)
         # Generate directories and configs
-        scale_factors: Dict[str, int] = {}
         slots: Dict[str, int] = {}
         for svc in BaseService.resolve(self.config.services):
             svc_cfg = self.config.services.get(svc.name)
-            # Process scale
-            scale = svc_cfg.scale if svc_cfg else 1
-            if scale <= 0:
-                continue  # Disabled
-            if svc.allow_scale:
-                scale_factors[svc.name] = scale
-            elif scale > 1:
-                logger.warn(
-                    "%s service is not supports scale, reverting to 1",
-                    svc.name,
-                )
             # Process slots
             if svc.require_slots:
-                slots[svc.name] = scale
+                slots[svc.name] = svc_cfg.scale if svc_cfg else 1
             # Create config
             svc.prepare_compose_config(self.config, svc_cfg)
             # Service discovery
             sd = svc.get_service_discovery(self.config, svc_cfg)
             if sd:
                 self._configure_service_discovery(svc.name, sd)
-        # Save scales
-        write_file(
-            etc / "scale.cfg",
-            "\n".join(f"{k}: {v}" for k, v in sorted(scale_factors.items())),
-            backup_path=etc / "scale.cfg.prev",
-        )
         # Save slots
         write_file(
             etc / "slots.cfg",
@@ -92,7 +74,7 @@ class ComposeTarget(BaseTarget):
     def _get_config_dict(self: "ComposeTarget") -> Dict[str, Any]:
         """Get dict of docker-compose.yml."""
         r = {
-            "version": "3",
+            "version": "3.3",
             "services": self._get_services_config(),
             "networks": {
                 "noc": {